Created Intro for Nuclio readme
Nuclio is a high-performance “serverless” framework that can compute intensive workloads. It is well integrated with popular data science tools and can support a variety of data and streaming sources.Nuclio can be configured as a standalone Docker container or on top of an existing Kubernetes or minikube.Once confiugured nuclio functions can be deployed over the CLI or throught the dashbaord that it provides.